frontend: router: Extract createRouteURL to own file
