+++ frontend/src/components/App/Home/ClusterContextMenu.tsx
@@ -28,7 +28,7 @@ import { useHistory } from 'react-router-dom';
 import helpers from '../../../helpers';
 import { deleteCluster } from '../../../lib/k8s/api/v1/clusterApi';
 import { Cluster } from '../../../lib/k8s/cluster';
-import { createRouteURL } from '../../../lib/router';
+import { createRouteURL } from '../../../lib/router/createRouteURL';
 import { useId } from '../../../lib/util';
 import { setConfig } from '../../../redux/configSlice';
 import { useTypedSelector } from '../../../redux/hooks';
