+++ frontend/src/components/App/Home/RecentClusters.tsx
@@ -26,7 +26,7 @@ import { isElectron } from '../../../helpers/isElectron';
 import { getRecentClusters, setRecentCluster } from '../../../helpers/recentClusters';
 import { formatClusterPathParam, getClusterPrefixedPath } from '../../../lib/cluster';
 import { Cluster } from '../../../lib/k8s/cluster';
-import { createRouteURL } from '../../../lib/router';
+import { createRouteURL } from '../../../lib/router/createRouteURL';
 import { MULTI_HOME_ENABLED } from './config';
 import SquareButton from './SquareButton';
 const ToggleButton = styled(MuiToggledButton)({
