+++ frontend/src/components/App/TopBar.tsx
@@ -33,7 +33,7 @@ import { useHistory } from 'react-router-dom';
 import { getProductName, getVersion } from '../../helpers/getProductInfo';
 import { logout } from '../../lib/auth';
 import { useCluster, useClustersConf } from '../../lib/k8s';
-import { createRouteURL } from '../../lib/router';
+import { createRouteURL } from '../../lib/router/createRouteURL';
 import {
   AppBarAction,
   AppBarActionsProcessor,
