+++ frontend/src/components/common/Link.tsx
@@ -21,8 +21,8 @@ import { Link as RouterLink } from 'react-router-dom';
 import { formatClusterPathParam, getCluster, getSelectedClusters } from '../../lib/cluster';
 import { kubeObjectQueryKey, useEndpoints } from '../../lib/k8s/api/v2/hooks';
 import type { KubeObject } from '../../lib/k8s/KubeObject';
-import type { RouteURLProps } from '../../lib/router';
-import { createRouteURL } from '../../lib/router';
+import type { RouteURLProps } from '../../lib/router/createRouteURL';
+import { createRouteURL } from '../../lib/router/createRouteURL';
 import { useTypedSelector } from '../../redux/hooks';
 import { Activity } from '../activity/Activity';
 import { canRenderDetails, KubeObjectDetails } from '../resourceMap/details/KubeNodeDetails';
