+++ frontend/src/components/common/Resource/Resource.tsx
@@ -42,7 +42,8 @@ import { KubeObjectInterface } from '../../../lib/k8s/KubeObject';
 import { KubeObjectClass } from '../../../lib/k8s/KubeObject';
 import Pod, { KubePod, KubeVolume } from '../../../lib/k8s/pod';
 import { METRIC_REFETCH_INTERVAL_MS, PodMetrics } from '../../../lib/k8s/PodMetrics';
-import { createRouteURL, RouteURLProps } from '../../../lib/router';
+import { RouteURLProps } from '../../../lib/router';
+import { createRouteURL } from '../../../lib/router/createRouteURL';
 import { getThemeName } from '../../../lib/themes';
 import { localeDate, useId } from '../../../lib/util';
 import { HeadlampEventType, useEventCallback } from '../../../redux/headlampEventSlice';
