+++ frontend/src/components/project/NewProjectPopup.tsx
@@ -37,7 +37,7 @@ import { apply } from '../../lib/k8s/api/v1/apply';
 import { ApiError } from '../../lib/k8s/api/v2/ApiError';
 import { KubeObjectInterface } from '../../lib/k8s/KubeObject';
 import Namespace from '../../lib/k8s/namespace';
-import { createRouteURL } from '../../lib/router';
+import { createRouteURL } from '../../lib/router/createRouteURL';
 import { useTypedSelector } from '../../redux/hooks';
 import { PROJECT_ID_LABEL, toKubernetesName } from './projectUtils';
 /**
